+++ include/cecxx/benchmark/detail/types.hpp
@@ -3,13 +3,16 @@
 #include <type_traits>
 #include <vector>
 
+#include "cecxx/types.hpp"
+
 namespace cecxx::benchmark::detail {
 
 template <typename Enum> constexpr auto to_underlying(Enum e) noexcept -> std::underlying_type_t<Enum> {
   return static_cast<std::underlying_type_t<Enum>>(e);
 }
 
-using fn_num = int;
+using fn_num = u8;
+
 template <typename T> using table_data = std::vector<T>;
 
 enum class table_type_t { rotate, shift, shuffle };
@@ -20,7 +23,7 @@ struct affine_mask_t {
   do_affine_trans rot{};
   do_affine_trans shift{};
   do_affine_trans shuffle{};
-  double rate{1.0};
+  f64 rate{1.0};
 };
 
 }  // namespace cecxx::benchmark::detail
